# backend/oauth/openapi.yaml
openapi: 3.1.0
info:
  title: ConstrSW - OAuth API Gateway
  description: API Gateway para o Keycloak, implementando os requisitos do T1.
  version: 1.0.0
servers:
  - url: http://localhost:8180
    description: Servidor de desenvolvimento (externo)
  - url: http://oauth:8000
    description: Servidor interno (Docker)
...
paths:
  /api/v1/login:
    post:
      tags:
        - Authentication
      summary: Autenticação de Usuário
      description: Autentica o usuário com username e password, usando o 'password' grant type do Keycloak. Retorna access_token e refresh_token.
      operationId: login
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  format: email
                  example: joao.silva@email.com
                password:
                  type: string
                  format: password
                  example: strongPassword123
      responses:
        '201':
          description: Autenticação bem-sucedida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Credenciais inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/refresh:
    post:
      tags:
        - Authentication
      summary: Renovação de Access Token
      description: Obtém um novo access_token e refresh_token usando um refresh_token válido.
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - refresh_token
              properties:
                refresh_token:
                  type: string
                  example: eyJhbGciOiJIUzI1NiIsIn...
      responses:
        '201':
          description: Token renovado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Refresh token inválido ou expirado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/validate:
    post:
      tags:
        - Authentication
      summary: Validação de Access Token
      description: Valida um access_token, verificando sua autenticidade e se não está expirado junto ao provedor de identidade (Keycloak).
      operationId: validateToken
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Token válido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntrospectResponse'
        '401':
          description: Token inválido ou expirado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtido através do endpoint /api/v1/login
  schemas:
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
          description: Token de acesso JWT
          example: eyJhbGciOiJSUzI1NiIsIn...
        expires_in:
          type: integer
          description: Tempo de expiração do access token em segundos
          example: 300
        refresh_expires_in:
          type: integer
          description: Tempo de expiração do refresh token em segundos
          example: 1800
        refresh_token:
          type: string
          description: Token de renovação
          example: eyJhbGciOiJIUzI1NiIsIn...
        token_type:
          type: string
          description: Tipo do token
          example: Bearer
    IntrospectResponse:
      type: object
      properties:
        active:
          type: boolean
          description: Indica se o token está ativo
          example: true
        username:
          type: string
          nullable: true
          description: Nome de usuário associado ao token
          example: joao.silva@email.com
        scope:
          type: string
          nullable: true
          description: Escopos do token
        client_id:
          type: string
          nullable: true
          description: ID do cliente
        exp:
          type: integer
          nullable: true
          description: Timestamp de expiração
        iat:
          type: integer
          nullable: true
          description: Timestamp de emissão
...
    Error:
      type: object
      properties:
        detail:
          type: string
          description: Mensagem de erro
          example: Credenciais inválidas. Verifique o usuário e a senha.